Add remaining subComponents for step 1.
