+++ src/components/registration/step1SubComponents/Email.js
@@ -1,0 +1,37 @@
+import React from 'react'
+
+const Email = ({ registrationForm, setRegistrationForm }) => {
+    return (
+      <Form.Item
+        label="Email"
+        rules={[
+          {
+            type: 'email',
+            message: 'Please enter a valid email address.',
+          },
+          {
+            required: true,
+            message: 'Email is required.',
+          },
+        ]}
+      >
+        <Input
+          name="email"
+          value={registrationForm.email}
+          onChange={(e) => {
+            const updatedRegForm = {
+              ...registrationForm,
+              email: e.target.value,
+            };
+            setRegistrationForm(updatedRegForm);
+          }}
+        />
+      </Form.Item>
+    )
+}
+
+const mapStateToProps = (state) => {
+  return { registrationForm: state.registrationForm };
+};
+
+export default connect(mapStateToProps, actions)(Email);
