+++ src/components/registration/step1SubComponents/Phone.js
@@ -1,0 +1,44 @@
+import React from 'react';
+
+const Phone= ({ registrationForm, setRegistrationForm }) => {
+    return (
+      <Form.Item
+        label={
+          <span>
+            Phone&nbsp;
+            <Tooltip title="USA & Canada country code is +1">
+              <QuestionCircleOutlined />
+            </Tooltip>
+          </span>
+        }
+        rules={[
+          {
+            required: true,
+            message: 'Phone number is required.',
+          },
+        ]}
+      >
+        <Input
+          name="phone"
+          style={{
+            width: '100%',
+          }}
+          placeholder="+1 (123) 123-4567"
+          value={registrationForm.phone}
+          onChange={(e) => {
+            const updatedRegForm = {
+              ...registrationForm,
+              phone: e.target.value,
+            };
+            setRegistrationForm(updatedRegForm);
+          }}
+        />
+      </Form.Item>
+    )
+}
+
+const mapStateToProps = (state) => {
+  return { registrationForm: state.registrationForm };
+};
+
+export default connect(mapStateToProps, actions)(Phone);
